# doc/api.yaml
openapi: '3.0.2'
info:
  title: WASAPhoto API Specifications
  description: |-
    Keep in touch with your friends by sharing photos of special moments, thanks to WASAPhoto! 
    You can upload your photos directly from your PC, and they will be visible to everyone following you.
  version: "1.0"
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: Bearer token for authentication in WASAPhoto.
      bearerFormat: '^[a-z0-9]{64}$'
  schemas:
    ID:
      title: ID
      description: |-
        Autoincrement integer field used for identify a particular resource (a comment, a post, ecc.)
      type: integer
      example: 795739
    Username:
      title: Username
      description: Username of a user
      type: string
      example: "chri_genna02"
      minLength: 8
      maxLength: 16
      pattern: '^[a-zA-Z0-9_-]{8,16}$'
    User:
      title: User
      description: |-
        A user in WASAPhoto is represented by a unique ID, its unique username, name and birthdate.
      properties:
        uuid:
          $ref: '#/components/schemas/ID'
        username:
          $ref: '#/components/schemas/Username'
        profilePic:
          $ref: "#/components/schemas/Photo"
        name:
          description: Real-life name/surname of the user
          type: string
          pattern: ''
          minLength: 2
          maxLength: 32
          example: Christian Gennarelli 
        birthdate:
          $ref: '#/components/schemas/Date'
...
    Photo:
      title: Photo
      format: binary
      type: string
      description: |-
        Binary representation of a photo, wrapped in a base64-encoded string. 
        Only multiple profile pics are sent via base64 for efficiency reasons.
        URL pointing to post images will be first sent, and then the client will have the ability
        to obtain the photo via a separate endpoint, i.e. GetPhotoFromURL.
      minLength: 0
      maxLength: 2000000 # 2MB photo
      example: iVBORw0KGgbW8gY2hyaXN0aWFuCghyaXN0aWFuCgw0KGgbW8gYyaXN0aWFuCgw0KGgbW8gW8gYyaXN0aW8gYyaXN0aWF
    PhotoURL:
      title: PhotoURL
      type: string
      description: |-
        URL of the photo, actually stored on server-side.
        The actual image will be sent as response of a separate request.
      minLength: 45 # http://WASAPhotoserver.com/photos/posts/1.png
      maxLength: 60 # http://WASAPhotoserver.com/photos/profile_pics/999999999.png
      example: "http://WASAPhotoserver.com/photos/profile_pics/default.png"
    Profile:
      title: Profile
      description: |-
        Profile of an user, made up of its informations and the its posts on WASASPhoto.
      properties:
        user:
          $ref: "#/components/schemas/User"
        posts:
          type: object
          description: |- 
            Collection of the posts posted by this user.
            At most 16 posts are returned at each request, according to some backend principle (example: datetime)
          items:
            $ref: '#/components/schemas/PostsStream'
          minItems: 0
          maxItems: 16
    Post:
      title: Post
      description: |-
        Post containing a photo, alongside the user that made it, its description and its date of creation.
      properties:
        post_id:
          $ref: '#/components/schemas/ID'
        photo: 
          $ref: "#/components/schemas/PhotoURL"
        creation-datetime: # Useful for the representation of the poststream, which must be displayed in reverse chronological order
          $ref: '#/components/schemas/Datetime'
        description:
          description: Description of the photo provided by the user (if existing).
          type: string
          minLength: 0
          maxLength: 128
          pattern: ''
    PostsStream:
      title: Stream
      description: |-
        Stream of posts.
        Only 16 posts are returned at each request.
      type: array
      items:
        $ref: '#/components/schemas/Post'
      minItems: 0
      maxItems: 16
...
    Date:
      title: date
      description: Date.
      example: "2023-11-29"
      type: string
      format: date
      pattern: '^([0-9]{4})-(0[1-9]|1[0-2])-(0[1-9]|[12][0-9]|3[01])$'
      minLength: 10
      maxLength: 10  
    Datetime:
      title: datetime
      description: |-
        Date and time.
      example: "2023-11-29 23:00:00"
      type: string
      format: date-time
      pattern: '^([0-9]{4})-(0[1-9]|1[0-2])-(0[1-9]|[12][0-9]|3[01]) (0[0-9]|1[0-9]|2[0-3]):([0-5][0-9]):([0-5][0-9])$'
      minLength: 19
      maxLength: 19  
    Error:
      title: Error
      description: |-
        Schema defining what is an error when returned in response of a HTTP requests.
      properties:
        code: # Code returned by the server
          type: string
          description: Status code returned directly by the server(s).
          example: "404"
          minLength: 3
          maxLength: 3
        description:  
          type: string
          description: |- 
            Description of the error in human-readable language.
          example: 'Username not found. Please check that the username is correct, and try again.'
          minLength: 1
          maxLength: 128
          pattern: ''
paths:
  /session:
    post:
      operationId: doLogin
      tags: ['LOGIN']
      summary: Login
      description: |-
        If the user does not exist, it will be created,
        and an identifier is returned.
        If the user exists, the user identifier is returned.
      requestBody:
        description: User details (here, just the username)
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Username'
        required: true
      responses:
        '201':
          description: User log-in action successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ID'
        '400': # Not acceptable
          description: Content not conformed to the criteria specified.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500': # Internal server error
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /users/{username}/profile/:
    parameters:
      - in: path
        name: username
        description: The user's username.
        schema:
          $ref: '#/components/schemas/Username'
        required: true
      
    get:
      operationId: getUserProfile
      tags: ['PROFILE']
      summary: See user profile
      description: |-
        Return all the informations about the specified user, 
        included its username, real name and date of birth.
      security:
        - BearerAuth: [] # We ask for the Auth token to check if the user has -been- banned
      responses:
        '200': # OK - Profile found
            description: Returned profile of the searched user.
            content:
              application/json:
                schema: 
                  $ref: '#/components/schemas/Profile'
        '204': # No content
          description: No corresponding profile was found.
        '400': # Bad request
          description: The request cannot be processed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401': # Unauthenticated
          description: User is not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403': # Unauthorized
          description: Authenticated user is not authorized to see the profile of the requested user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404': # Profile not found
          description: |-
              No user has been found with the provided username, 
              thus no information can be returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500': # Internal server error.
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    put:
      operationId: setMyUserName
      tags: ['PROFILE']
      summary: Update username
      description: |-
        A user can change its username with a new one
        (provided that the new one is valid and not already used).
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Username'
      responses: 
        '201': # OK - Username updated
          description: New username of this user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Username'
        '400': # Bad request
          description: Bad request provided.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401': # Unauthenticated
          description: The client has not been authenticated (yet).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403': # Unauthorized
          description: |-
            The authenticated user cannot change the username of another user.
            That is, the authenticated username and the one provided in the path do NOT coincide.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404': # Profile not found
          description: |-
              No user has been found with the provided username, 
              thus no information can be returned. 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '406': # New username provided already in use
          description: |- 
            The "new_username" provided has been found to be already in use in the system.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500': # Internal server error.
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /photos/{photo_path}:
    parameters:
      - in: path
        name: photo_path
        description: |-
          URL path of the requested photo.
        schema:
          type: string
          minLength: 11 # 6 (posts/) + 4 (x.png) -> posts/1.png
          maxLength: 33 # 13 ("profile_pics/") + 16 (max username length) -> profile_pics/wasaphotoexample.png
          pattern: ''
          example: "profile_pics/wasaphotoexample.png"
        required: true
    
    get:
      operationId: getPhotoFromURL
      tags: ['PHOTO']
      summary: Get the requested photo
      description: |-
        Given the URL path to the server resource representing the requested photo,
        sent it to the requesting client.
      responses:
        '200': # OK 
          description: The photo was found, thus sent to the client.
          content:
            image/png:
              schema: 
                description: Requested photo (in PNG).
                type: string
                format: binary
        '400': # Bad request
          description: Provided path not valid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404': # Photo not found
          description: Requested photo has not been found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500': # Internal server error
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'